+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Fixed
 
 - Correction for `InvalidDestinationId` exception arguments in destination handler.
+- Destination handler now only checks entity ID values when checking inserted packets
 
 # [v0.4.0] 2024-11-08